chore: rename `nfcabortconfirmation` package to `confirmabort`
Resolves: DCMAW-8099
