+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/confirmabort/ConfirmNoChippedID.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.nfcabortconfirmation
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.confirmabort
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -7,16 +7,16 @@ import uk.gov.android.ui.theme.m3.GdsTheme
 import uk.gov.onelogin.criorchestrator.libraries.composeutils.LightDarkBothLocalesPreview
 
 @Composable
-internal fun ConfirmNoNonChippedID(modifier: Modifier = Modifier) {
+internal fun ConfirmNoChippedID(modifier: Modifier = Modifier) {
     Text(
-        text = "DCMAW-10678 | Android | Document Selection | No NFC enabled Abort Confirmation Screen",
+        text = "DCMAW-10691 | Android | Document Selection | NFC enabled Abort Confirmation Screen",
         modifier = modifier,
     )
 }
 
 @LightDarkBothLocalesPreview
 @Composable
-internal fun PreviewConfirmNoNonChippedID() {
+internal fun PreviewConfirmNoChippedID() {
     GdsTheme {
         ConfirmNoChippedID()
     }
